+++ WorldServer/Entities/CMgrAvatorQuestHistory.cpp
@@ -95,7 +95,7 @@ int CMgrAvatorQuestHistory::GetTotalWaitSize()
 
 void CMgrAvatorQuestHistory::WriteFile(char *pszFileName, char *pszLog)
 {
-  const unsigned int size = std::strlen(pszLog);
+  const unsigned int size = static_cast<unsigned int>(std::strlen(pszLog));
   unsigned int outIndex[4]{};
   if (size < 0x12C && m_listLogDataEmpty.PopNode_Front(outIndex))
   {
